fix predict errors with GCC 8.3

+++ src/predict/mepo.cpp
@@ -56,7 +56,7 @@ MePo::MePo(LVecVec deps, LVecVec syms, LVecVec sym_ths, long sym_num)
 }
 
 void MePo::learn(long from, long to) {
-  for (unsigned i = from; i < to; ++i)
+  for (long i = from; i < to; ++i)
     tfidf.add(syms[i]);
 }
